+++ .github/agents/new-unit.agent.md
@@ -56,7 +56,7 @@ public readonly struct {Name} : ISiUnit, I{Dimension}
 
 ### 2. Metric Unit (`IMetricUnit`)
 
-**When**: The unit is accepted by the SI system but is not itself an SI unit. Metric units support SI metric prefixes (`Kilo`, `Mega`, `Milli`, etc.).
+**When**: The unit is accepted by the SI system but is not itself an SI unit. This includes all units listed in the SI Brochure as ["non-SI units accepted for use with SI"](https://en.wikipedia.org/wiki/International_System_of_Units#Non-SI_units_accepted_for_use_with_SI) (e.g. minute, hour, day, degree, arcminute, arcsecond, astronomical unit, hectare, litre, tonne, dalton, electronvolt). Metric units support SI metric prefixes (`Kilo`, `Mega`, `Milli`, etc.).
 
 **Key property**: Must implement `ITransform` via `ToSi(Transformation)`. Support metric prefixes by default.
 
@@ -80,7 +80,7 @@ public readonly struct {Name} : IMetricUnit, I{Dimension}
 }
 ```
 
-**Examples**: `Hour`, `Minute`, `Gram`, `Tonne`, `Day`, `Week`, `Are`, `Bar`, `AstronomicalUnit`, `Ångström`, `HorsePower`, `Bit`, `Byte`
+**Examples**: `Hour`, `Minute`, `Gram`, `Tonne`, `Day`, `Week`, `Are`, `Bar`, `AstronomicalUnit`, `Ångström`, `Degree`, `HorsePower`, `Bit`, `Byte`
 
 ---
